==> 03-cleaning.Rmd <==
# Data Transformation

## Import Packages and Dataset

```{r include=FALSE}
knitr::write_bib("knitr", file = "packages.bib")
```

IPUMSR is a powerful package developed to "import census, survey and geographic data provided by 'IPUMS' into R plus tools to help use the associated metadata to make analysis easier" ^[Ellis, Burk, Grover, Minnesota Population Center (2017). ipumsr: Read 'IPUMS' Extract Files. R package version 0.4.5. https://cran.r-project.org/web/packages/ipumsr/index.html]. 

Here, we use the function `read_ipums_ddi` to read in the data codebook, which is downloaded along with the extract and stored in XML format. In an R development environment such as RStudio, you can use the command `ipums_view(ddi)` to open an interactive codebook through the built-in viewer. 

Then, we can use `read_ipums_micro(ddi)` to load the actual dataset into the environment, given data file is placed next the XML codebook. 

```{r echo=TRUE}
library(ipumsr) 

# Read in codebook
ddi <- read_ipums_ddi("./data/usa_00038.xml")
# Load in data
data <- read_ipums_micro(ddi)
```

## Reduce Cardinality and Decode Data

As discussed in the previous chapter, one issue with the data is that some variables have a high cardinality with roughly even proportion. For these categories, we cannot lump the minority values into the Other category. For this project, we collect the multi-layer categorization from the IPUMS documentation to cluster the occupation and industry into larger categories ^[https://usa.ipums.org/usa-action/variables/OCC1950#codes_section]. We also make a decision to use `lbl_na_if` function from IPUMSR to mark non-occupational responses such as "keeping house" as missing values. 

```{r  include=FALSE}

library(tidyverse) 

# Remove varialbes with lower usability, as explained in previous chapter
data <- data %>% select(-c(SERIAL, PERNUM, RELATED, RACED, BPLD, MBPLD, FBPLD, YRSUSA2, MTONGUE, MTONGUED, PRESGL))

# Data augmentation for categories of occupations and industries
occ_cat <- read_csv("./data/occupation_categories.csv")
ind_cat <- read_csv("./data/industry_categories.csv")
data <- left_join(data, occ_cat, by ='OCC1950')
data <- left_join(data, ind_cat, by ='IND1950')

# NA relabeling based on code range
data$OCC1950 <- lbl_na_if(data$OCC1950, ~.val > 970) %>% as_factor()
data$IND1950 <- lbl_na_if(data$IND1950, ~.val >= 975 | .val == 0) %>% as_factor()
data$OCCSCORE <- as.numeric(data$OCCSCORE)

# Factorize columns
df <- as_factor(data)

```

## Handling N/A, Special Cases and Enforce Data Types

The other data issue (tail values) is addressed through a series of column-specific data transformations. More specifically, we handle different codes for N/A, collapse tail values into larger categories, and ensure that all values are in the correct data type for that column. When the meaning is clear, we also trim the value to make it concise. The resulting dataset is saved as a CSV file for downstream analysis. For the assumptions and actual processes, please refer to the codes and comments in the Rmarkdown source.

```{r  include=FALSE}

# Customized cleaning for each column

# YEAR
df$YEAR <- as.numeric(as.character(df$YEAR))
# CITY
# FAMSIZE
df$FAMSIZE <- as.character(df$FAMSIZE)
df$FAMSIZE[df$FAMSIZE == "1 family member present"] = "1"
df$FAMSIZE[df$FAMSIZE == "2 family members present"] = "2"
df$FAMSIZE <- as.numeric(df$FAMSIZE)
# RELATE
df$RELATE <- as.character(df$RELATE)
df$RELATE[df$RELATE == "Head/Householder"] = "Head"
df$RELATE[df$RELATE == "Child-in-law"] = "Child"
df$RELATE[df$RELATE == "Grandchild"] = "Child"
df$RELATE[df$RELATE == "Parent-in-Law"] = "Parent"
df$RELATE[df$RELATE == "Sibling-in-Law"] = "Sibling"
df$RELATE[df$RELATE == "Other relatives"] = "Relative"
df$RELATE[df$RELATE == "Partner, friend, visitor"] = "Relative"
df$RELATE[df$RELATE == "Other non-relatives"] = "Other"
df$RELATE[df$RELATE == "Institutional inmates"] = "Inmate"
df$RELATE <- as.factor(df$RELATE)
# SEX
# AGE
df$AGE <- as.character(df$AGE)
df$AGE[df$AGE == "Less than 1 year old"] = "0"
df$AGE[df$AGE == "90 (90+ in 1980 and 1990)"] = "90"
df$AGE[df$AGE %in% c("100 (100+ in 1960-1970)", "112 (112+ in the 1980 internal data)", "115 (115+ in the 1990 internal data)")] = "100"
df$AGE <- as.numeric(df$AGE)
# MARST
df$MARST <- as.character(df$MARST)
df$MARST[df$MARST == 'Married, spouse present'] = 'Married'
df$MARST[df$MARST == 'Never married/single'] = 'Single'
df$MARST[df$MARST %in% c('Married, spouse absent','Separated','Divorced','Widowed')] = 'Widowed or Divorced'
df$MARST <- as.character(df$MARST)
df$MARST <- as.factor(df$MARST)
# RACE
df$RACE <- as.character(df$RACE)
df$RACE[df$RACE == 'Black/African American/Negro'] = 'Black'
df$RACE[!(df$RACE %in% c('White','Black'))] = 'Other'
df$RACE <- as.factor(df$RACE)
# BPL
# MBPL
# FBPL
# NATIVITY
df$NATIVITY <- as.character(df$NATIVITY)
df$NATIVITY[df$NATIVITY %in% c("Native born, and both parents native born","Native born, and father foreign, mother native","Native born, and mother foreign, father native")] = 'Native born with native parent'
df$NATIVITY[df$NATIVITY == 'Native born, and both parents foreign'] = 'Native born with foreign parent'
df$NATIVITY[df$NATIVITY == 'N/A'] = NA
df$NATIVITY <- as.factor(df$NATIVITY)
# CITIZEN
df$CITIZEN <- as.character(df$CITIZEN)
df$CITIZEN[df$CITIZEN == 'Not a citizen, but has received first papers'] = 'Applying citizenship'
df$CITIZEN[!(df$CITIZEN %in% c('Naturalized citizen','Not a citizen', 'Applying citizenship'))] = NA
df$CITIZEN <- as.factor(df$CITIZEN)
# YRIMMIG
df$YRIMMIG <- as.character(df$YRIMMIG)
df$YRIMMIG = str_replace(df$YRIMMIG, " \\s*\\([^\\)]+\\)", "")
df$YRIMMIG[df$YRIMMIG %in% c('N/A', 'Not reported')] = NA
df$YRIMMIG <- as.numeric(df$YRIMMIG)
# Creating Years spent in US variable
df <- df %>%
 mutate(
  YRINUSA = YEAR - YRIMMIG
 )
# SCHOOL
df$SCHOOL <- as.character(df$SCHOOL)
df$SCHOOL[df$SCHOOL %in% c("N/A", "Unknown", "Missing")] = NA
df$SCHOOL <- as.factor(df$SCHOOL)
# LIT
df$LIT <- as.character(df$LIT)
df$LIT[df$LIT %in% c("Yes, literate (reads and writes)","Cannot read, can write","Cannot write, can read")] = 'Yes, literate'
df$LIT[df$LIT == 'No, illiterate (cannot read or write)'] = 'No, illiterate'
df$LIT[df$LIT %in% c("N/A", "Unknown, illegible or blank")] = NA
df$LIT <- as.factor(df$LIT)
# LABFORCE
df$LABFORCE <- as.character(df$LABFORCE)
df$LABFORCE[df$LABFORCE == 'N/A'] = NA
df$LABFORCE <- as.factor(df$LABFORCE)
# OCC1950
# IND1950
# INCWAGE
df$INCWAGE <- as.character(df$INCWAGE)
df$INCWAGE[df$INCWAGE %in% c('N/A','Missing')] = NA
df$INCWAGE <- as.numeric(df$INCWAGE)
# OCCSCORE
df$OCCSCORE[df$OCCSCORE == 0] = NA
# EDSCOR50
df$EDSCOR50 <- as.character(df$EDSCOR50)
df$EDSCOR50[df$EDSCOR50 %in% c('N/A')] = NA
df$EDSCOR50 <- as.numeric(df$EDSCOR50)

# Output data
write.csv(df, file = './data/ipums.csv', row.names = FALSE)
```
