==> 05-results.Rmd <==
# Results

```{r}
library('tidyverse')
```

```{r}
df <- read.csv(file = 'data/ipums.csv')
drop_columns = c('BPL', 'MBPL', 'FBPL', 'NATIVITY', 'CITIZEN', 'YRIMMIG', 
                 'YRINUSA', 'INCWAGE', 'MARST', 'OCC1950', 'IND1950', 'CITY')
data <- df %>%
  select(-one_of(drop_columns)) %>%
  filter(AGE>=18 & AGE<=66)
```

Before analysis, we first do data cleaning and select several features and data for analyzing. The following data contains features including three main categories, which are minority features, education features, and employment features.

1. Minority Feature: Respondents Age, Respondents Race
2. Education Feature: LIT(Whether could read and write)
3. Employment Feature: LABFORCE(Whether in the job market), OCCSCORE(Occupational income score), OCC1950CAT(Occupational category in 1950 standard), IND19050CAT(Industry category in 1950 standard)

The following analysis will give a glimpse of the minority working conditions, the influence of education levels, and the working conditions transition across different years.

## Working Conditions

### Gender and Employment

1. Labor force percentage for each gender: From the following figure, we could conclude that most of the male are in the labor force, while half of the female are not in the labor force. There is significant difference in working conditions between male and female.

```{r}
ggplot(data, aes(LABFORCE, group = SEX)) + 
  geom_bar(aes(y = ..prop.., fill = LABFORCE), stat="count") + 
  scale_y_continuous(labels=scales::percent) +
  facet_wrap(~SEX) +
  coord_flip() +
  xlab("Labor Force") + 
  ylab("Percentage in Each Gender") +
  ggtitle("Labor Force Percentage for each Gender") +
  theme(plot.title = element_text(hjust = 0.5)) 
```

2. Occupation score for each gender: The occupation score represents the salary for the workers. As shown in the following figure, the occupation score for male workers are significant higher than the female workers.

```{r}
ggplot(data, aes(x=SEX, y=OCCSCORE)) +
  geom_boxplot() +
  coord_flip() +
  xlab("Sex") +
  ylab("Occupation Score") +
  ggtitle("Occupation Score for each Gender") +
  theme(plot.title = element_text(hjust = 0.5)) 
```

3. Occupation category for each gender: There is difference between each gender of the occupation categories. Female tends to work as operatives, clerical and kindred and private household service workers. Male tends to work as craftsmen, managers and laborers.

```{r}
ggplot(data) +
  geom_bar(aes(x=fct_infreq(OCC1950CAT), fill=SEX), stat = "count") +
  xlab("Occupation Category") +
  ylab("Count of Records") + 
  ggtitle("Occupation Category for each Gender") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

4. Industry category for each gender: There is difference between male and female in industry categories. Female workers tend to work in manufacturing-nondurable goods, personal services, and professional and related services. Male workers tend to work in durable goods manufacturing, construction, and transportation.

```{r}
ggplot(data) +
  geom_bar(aes(x=fct_infreq(IND1950CAT), fill=SEX), stat = "count") +
  xlab("Industry Category") +
  ylab("Count of Records") + 
  ggtitle("Industry Category for each Gender") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

### Race and Employment

1. Labor force percentage for each Race: As shown in the following figure, there is little difference between each race labor force percentage. Black people labor force percentage is litter higher than white people.

```{r}
ggplot(data, aes(fct_infreq(LABFORCE), group = RACE)) + 
  geom_bar(aes(y = ..prop.., fill = LABFORCE), stat="count") + 
  scale_y_continuous(labels=scales::percent) +
  facet_wrap(~RACE) +
  xlab("Labor Force") + 
  ylab("Percentage in Each Subgroup") +
  ggtitle("Labor Force Percentage for each Race") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

2. Occupation Score for each Race: Although there is little difference between each race labor force percentage, there is significant difference between occupation score. It shows white people have significantly higher salary than black people and other races. And the variance in salary for white people is lees than other races. Therefore, it means salary for white people is high and centered.

```{r}
ggplot(data, aes(x=RACE, y=OCCSCORE)) +
  geom_boxplot() +
  coord_flip() + 
  ggtitle("Occupation Score for each Race") +
  theme(plot.title = element_text(hjust = 0.5)) 
```

3. Occupation category for each Race: Black people tend to work as operatives, service workers and laborers. White people tend to work as craftsmen, clerical and kindred, managers and sales workers.

```{r}
ggplot(data) +
  geom_bar(aes(x=fct_infreq(OCC1950CAT), fill=RACE), stat = "count") +
  xlab("Occupation Category") +
  ylab("Count of Records") + 
  ggtitle("Occupation Category for each Race") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

4. Industry category for each race: Black people tend to work in personal services industry.

```{r}
ggplot(data) +
  geom_bar(aes(x=fct_infreq(IND1950CAT), fill=RACE), stat = "count") +
  xlab("Industry Category") +
  ylab("Count of Records") + 
  ggtitle("Industry Category for each Race") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

## Education and Employment

1. Labor force for each gender and literate: As shown in the figure, there is little difference between those who are literate and who are illiterate in male workers. Female workers who are literate are more likely in the labor force.

```{r}
ggplot(data, aes(LABFORCE, group = LIT)) + 
  geom_bar(aes(y = ..prop.., fill = LABFORCE), stat="count") + 
  scale_y_continuous(labels=scales::percent) +
  facet_wrap(~SEX+LIT) +
  coord_flip() +
  xlab("Labor force") + 
  ylab("Percentage in Each Subgroup") +
  ggtitle("Labor Force Percentage for each Gender and Literate") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

2. Labor force for each race and literate: From the following figure, the people who are literate tend to be in the labor force. Especially for white people, white people who are literate significantly tend to work in the labor force.

```{r}
ggplot(data, aes(LABFORCE, group = LIT)) + 
  geom_bar(aes(y = ..prop.., fill = LABFORCE), stat="count") + 
  scale_y_continuous(labels=scales::percent) +
  facet_wrap(~RACE+LIT) +
  coord_flip() +
  xlab("Labor Force") + 
  ylab("Percentage in Each Subgroup") +
  ggtitle("Labor Force Percentage for each Race and Literate") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

3. Occupation score for each race and literate: From the following figure, we conclude that people who are literate tend to have higher salary than those who are not literate. Especially for female workers, literate female workers median salary are significantly highers than illiterate female workers.

```{r}
ggplot(data, aes(x=LIT, y=OCCSCORE)) +
  geom_boxplot() +
  facet_wrap(~SEX) +
  ggtitle("Occupation Score for each Race and Literate") +
  theme(plot.title = element_text(hjust = 0.5)) 
```

4. Occupation score for each race and literate: The figure shows literate people tend to have higher salary. Especially for black people, the median salary for literate black workers are higher than illiterate black workers.

```{r}
ggplot(data, aes(x=LIT, y=OCCSCORE)) +
  geom_boxplot() +
  facet_wrap(~RACE) +
  ggtitle("Occupation Score for each Race and Literate") +
  theme(plot.title = element_text(hjust = 0.5)) 
```

5. Occupation category and literate level: The following figure shows illiterate people tend to work as operatives, craftsmen and laborers, and literate people tend to work as clerical and kindred, sales workers, and professionals.

```{r}
ggplot(data) +
  geom_bar(aes(x=fct_infreq(OCC1950CAT), fill=LIT), stat = "count") +
  xlab("Occupation Category") +
  ylab("Count of Records") + 
  ggtitle("Occupation Category and Literate Level") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

6. Industry category and literate level: illiterate people tend to work in retail, manufacuring-nondurable goods, and personal services area. And literate people can work in industries like finance, business, public administration, etc.

```{r}
ggplot(data) +
  geom_bar(aes(x=fct_infreq(IND1950CAT), fill=LIT), stat = "count") +
  xlab("Industry Category") +
  ylab("Count of Records") + 
  ggtitle("Industry Category and Literate Level") +
  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) 
```

## Time series and Employment

1. Labor force for races and genders from 1850 to 1940: From the following figure, we could conclude that from 1850 to 1940, more and more female white people enter the labor force. The male workers start to enter the labor force from 1850 to 1920, and then some of them quit the labor force in 1930s and 1940s.

```{r}
ggplot(data, aes(x=YEAR, group = LABFORCE, fill=LABFORCE)) + 
  geom_bar(aes(y = ..prop.., fill = LABFORCE), stat="count", position = "fill") + 
  scale_y_continuous(labels=scales::percent) +
  facet_wrap(~SEX+RACE) +
  coord_flip() +
  scale_x_reverse(breaks = seq(1850, 1940, by = 10)) +
  ggtitle("Labor Force for Races and Genders from 1850 to 1940") +
  theme(legend.position="bottom", plot.title = element_text(hjust = 0.5))
```

2. Occupation score for races and genders from 1850 to 1940: As shown in the figure, from 1850 to 1940, the female occupation score increases, especially for white people. And the black male workers salary increases a little. The salary for white male is steady across different years.

```{r}
time_series <- data %>%
  group_by(YEAR, SEX, RACE) %>%
  summarise(
    OCCSCORE = mean(OCCSCORE, na.rm = TRUE)
  )
ggplot(time_series, aes(x=YEAR, y=OCCSCORE)) +
  geom_line(aes(color=RACE)) +
  scale_x_continuous(breaks = seq(1850, 1940, by = 10)) +
  theme(legend.position="bottom") + 
  facet_wrap(~SEX) +
  ggtitle("Occupation Score for Races and Genders from 1850 to 1940") +
  theme(legend.position="bottom", plot.title = element_text(hjust = 0.5))
```

## Conclusion

1. The minority working conditions are worse than the majority. Female workers labor force percentage is significantly less than the male workers, and the salary for female workers is less than male workers. There is little difference between difference races labor force percentage. However, the salary for the black workers is less than white workers.
2. There are significant difference in industry categories and types of work between male workers and female workers. Female workers tend to work in the industry like personal services as services workers, while male workers tend to work in construction, transportation as craftsman and laborer. For different races, black people tend to work in personal services as services workers, while white people tend to work in officials, finance industries as managers or professionals.
3. Education levels significantly influence the working conditions. Literate people tend to have higher salary and work at officials, business, finance area as professionals or managers. Illiterate people have lower salary, and work at construction, manufacturing as workers and craftsman.
4. From 1850 to 1940, minority people working conditions improve a lot. The salary and percentage of people in the labor force increase for female workers and black workers.
