==> 07-conclusion.Rmd <==
# Conclusion

## Summary

1. Minority Working Condition: The minority working conditions are worse than the majority. For difference genders, less female workers are in the labor force than male workers, and the salary for female workers is less than male workers. Female workers tend to work in the industry like personal services as services workers, while male workers tend to work in construction, transportation as craftsman and laborer. For different races, the salary for black workers is less than white workers. Black people tend to work in personal services as services workers, while white people tend to work in officials, finance industries as managers or professionals.
2. Influence of Education: Education levels significantly influence the working conditions. Literate people have higher salary than illiterate people, especially for white workers and female workers. Besides, literate people work in more sophisticated areas as "white collars".
3. Transition across Time: From 1850 to 1940, minority people working conditions improve a lot.

## Suggestion

1. Pay Attention to the Minority: Make policies from the government to pay more attention to minority workers, such as encourage companies to hire more minority workers, set up minimum salary standard for minority workers, etc.
2. Improve the Education Level: Encourage people to participate in education, especially for black people and female people. Schools admit more minority students. 

## Look Forward

1. Working Conditions Across Regions: Future study should pay attention to minority working conditions in different regions, since this study only include the samples from New York City.
2. More Influencing Factors: Future study should pay attention to other influencing factors such as family structures, geographical features, etc.
